requirements.txt
contourpy==1.2.1
CTkMessagebox==2.5
CTkTable==1.1
CTkToolTip==0.8
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.52.4
fuzzywuzzy==0.18.0
kiwisolver==1.4.5
Levenshtein==0.25.1
lock==2018.3.25.2110
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-immutable==1.1.1
python-Levenshtein==0.25.1
rapidfuzz==3.9.3
read-roi==1.6.0
six==1.16.0
typing_extensions==4.12.1
